Add gathering of Homebrew package versions to build
It's currently kind of a pain to assemble the versions of all the
installed Homebrew packages for an image. This is easy to generate, we
just needed a way to get it off the machine.

Add a step in the 'homebrew' role to generate a file with the info. Then
add a step to the Packer build that downloads that file locally into a
records directory that we will track in Git so we have a record of what
is installed on the various images.