Gather installed simulators during Mac image build
This has two effects: it gives us info to use in documentation for new
images, and it speeds up this command in the future when it's run during
builds. The first invocation of `xcrun simctl list` can take over 2
minutes now, so paying that cost during image building instead of during
jobs is preferable.